[WIP][TASK] Implement interfaces, base and refrence implementation
This change adds the interfaces for the CMS Health Check objects
along with a simple base implementation.

The PSR like implementation boils down to only two interfaces:

* `GlobalCheckInterface` as the outer main object handling
* `CheckInterface` is the interface for a single check

On top, a enhanced serializable reference implementation is added
for showcasing purpose during the great CloudFest Hackathon 2024.
The reference implementation may be extracted later into it's own
reference implementation package but kept for now to pave the path
for a quick POC implementation during the Hackathon.

Basic `PHPUnit` based unit testing is added along the way together
with a simple `php-cs-fixer` based cgl setup. Usually, that should
be provided in dedicated changes.